+++ dev-python/bpy/bpy-4.3.2-r4.ebuild
@@ -3,33 +3,54 @@
 
 EAPI=8
 
-PYTHON_COMPAT=( python3_13 )
+PYTHON_COMPAT=( python3_{11..12} )
 LLVM_COMPAT=( {18..19} )
 LLVM_OPTIONAL=1
 EGIT_LFS="yes"
 ROCM_VERSION="6.3"
 
-inherit rocm git-r3 check-reqs cmake cuda flag-o-matic python-single-r1 toolchain-funcs llvm-r1
+inherit ffmpeg-compat rocm check-reqs cmake cuda flag-o-matic python-single-r1 toolchain-funcs llvm-r1
 
 DESCRIPTION="3D Creation/Animation/Publishing System"
 HOMEPAGE="https://www.blender.org"
 LICENSE="GPL-3+ cycles? ( Apache-2.0 )"
 SLOT="0/$(ver_cut 1-2)"
 
-EGIT_REPO_URI="https://projects.blender.org/blender/blender.git"
-ASSETS_EGIT_REPO_URI="https://projects.blender.org/blender/blender-assets.git"
-ASSETS_EGIT_LOCAL_ID="${CATEGORY}/${PN}/${SLOT%/*}-assets"
+HAS_ASSETS=1
+HAS_ADDONS=0
+HAS_RELEASED=$([[ ${PV} != *9999* && ${PV} != *_beta* ]] && echo 1 || echo 0)
 
-if [[ ${PV} != *9999* ]]; then
+if [ "${HAS_RELEASED}" -eq 1 ]; then
+	IS_LIVE=0
+	SRC_URI="
+		https://projects.blender.org/blender/blender/archive/v${PV}.tar.gz -> ${P}.tar.gz
+	"
+	S="${WORKDIR}/${PN}"
+
+	if [ "${HAS_ASSETS}" -eq 1 ]; then
+		SRC_URI+="
+			https://projects.blender.org/blender/blender-assets/archive/v${PV}.tar.gz -> ${P}-assets.tar.gz
+		"
+	fi
+
+	if [ "${HAS_ADDONS}" -eq 1 ]; then
+		SRC_URI+="
+			https://projects.blender.org/blender/blender-addons/archive/v${PV}.tar.gz -> ${P}-addons.tar.gz
+		"
+	fi
+
+	KEYWORDS="~amd64"
+else
+	inherit git-r3
+	EGIT_REPO_URI="https://projects.blender.org/blender/blender.git"
+	ASSETS_EGIT_REPO_URI="https://projects.blender.org/blender/blender-assets.git"
+	ADDONS_EGIT_REPO_URI="https://projects.blender.org/blender/blender-addons.git"
 	if [[ ${PV} != *_beta* ]]; then
-		EGIT_COMMIT="v${PV}"
+		EGIT_BRANCH="main"
 	else
 		EGIT_BRANCH="blender-v$(ver_cut 1-2)-release"
 	fi
-	ASSETS_EGIT_BRANCH="${EGIT_BRANCH}"
-	KEYWORDS="~amd64"
-else
-	ASSETS_EGIT_BRANCH="main"
+
 	# special branches
 	if [[ ${PV} == *99991* ]]; then
 		EGIT_BRANCH="npr-prototype"
@@ -75,7 +96,7 @@ RDEPEND="${PYTHON_DEPS}
 	media-libs/libjpeg-turbo:=
 	media-libs/libpng:=
 	media-libs/libsamplerate
-	>=media-libs/openimageio-2.5.6.0:=
+	>=media-libs/openimageio-2.4.6.0:=
 	sys-libs/zlib:=
 	virtual/glu
 	virtual/libintl
@@ -84,10 +105,9 @@ RDEPEND="${PYTHON_DEPS}
 	collada? ( >=media-libs/opencollada-1.6.68 )
 	color-management? ( media-libs/opencolorio:= )
 	cuda? ( dev-util/nvidia-cuda-toolkit:= )
-	embree? ( media-libs/embree:=[raymask] )
+	embree? ( >=media-libs/embree-3.13.0:=[raymask] )
 	ffmpeg? (
-		media-video/ffmpeg:=[encode(+),jpeg2k?,opus,theora,vorbis,vpx,x264,xvid]
-		|| ( media-video/ffmpeg[lame(-)] media-video/ffmpeg[mp3(-)] )
+		media-video/ffmpeg-compat:6=[encode(+),lame,jpeg2k?,opus,theora,vorbis,vpx,x264,xvid]
 	)
 	fftw? ( sci-libs/fftw:3.0= )
 	gmp? ( dev-libs/gmp[cxx] )
@@ -102,13 +122,13 @@ RDEPEND="${PYTHON_DEPS}
 	openal? ( media-libs/openal )
 	oidn? ( >=media-libs/oidn-2.3.2:= )
 	openexr? (
-		>=dev-libs/imath-3.1.7:=
-		>=media-libs/openexr-3.2.1:0=
+		>=dev-libs/imath-3.1.4-r2:=
+		>=media-libs/openexr-3:0=
 	)
 	openpgl? ( media-libs/openpgl:= )
 	opensubdiv? ( >=media-libs/opensubdiv-3.6.0-r2[opengl,cuda?,openmp?,tbb?] )
 	openvdb? (
-		>=media-gfx/openvdb-11.0.0:=[nanovdb?]
+		>=media-gfx/openvdb-10.1.0:=[nanovdb?]
 		dev-libs/c-blosc:=
 	)
 	optix? ( dev-libs/optix )
@@ -150,7 +170,7 @@ RDEPEND="${PYTHON_DEPS}
 		x11-libs/libXi
 		x11-libs/libXxf86vm
 	)
-	hiprt? ( dev-libs/hiprt:2.5=[${LLVM_USEDEP}] )
+	hiprt? ( dev-libs/hiprt:2.3= )
 "
 
 DEPEND="${RDEPEND}
@@ -176,11 +196,11 @@ BDEPEND="
 "
 
 PATCHES=(
+	"${FILESDIR}/${PN}-4.1.1-openvdb-11.patch"
 	"${FILESDIR}/${PN}-4.1.1-clang.patch"
-	"${FILESDIR}/${PN}-4.4.0-hiprt-parallel.patch"
+	"${FILESDIR}/${PN}-4.3.2-hiprt-parallel.patch"
 	"${FILESDIR}/${PN}-4.3.2-hipcc-path.patch"
-	"${FILESDIR}/${PN}-4.4.0-cycles-runtime-path.patch"
-	"${FILESDIR}/${PN}-4.4.0-functional-header.patch"
+	"${FILESDIR}/${PN}-4.3.2-cycles-runtime-path.patch"
 )
 
 blender_check_requirements() {
@@ -226,9 +246,25 @@ pkg_setup() {
 }
 
 src_unpack() {
-	git-r3_fetch "${ASSETS_EGIT_REPO_URI}" "${ASSETS_EGIT_BRANCH}" "${ASSETS_EGIT_LOCAL_ID}"
-	git-r3_checkout "${ASSETS_EGIT_REPO_URI}" "${WORKDIR}/blender-assets" "${ASSETS_EGIT_LOCAL_ID}"
-	git-r3_src_unpack
+	if [ "${HAS_RELEASED}" -eq 1 ]; then
+		default
+
+		if [ "${HAS_ADDONS}" -eq 1 ]; then
+			mv "${WORKDIR}/blender-addons" "${S}/scripts/addons"
+		fi
+	else
+		if [ "${HAS_ASSETS}" -eq 1 ]; then
+			git-r3_fetch "${ASSETS_EGIT_REPO_URI}" "${EGIT_BRANCH}"
+			git-r3_checkout "${ASSETS_EGIT_REPO_URI}" "${WORKDIR}/blender-assets"
+		fi
+
+		if [ "${HAS_ADDONS}" -eq 1 ]; then
+			git-r3_fetch "${ADDONS_EGIT_REPO_URI}" "${EGIT_BRANCH}"
+			git-r3_checkout "${ADDONS_EGIT_LOCAL_ID}" "${S}/scripts/addons"
+		fi
+
+		git-r3_src_unpack
+	fi
 }
 
 src_prepare() {
@@ -245,9 +281,11 @@ src_prepare() {
 		sed -e "s/extern_vulkan_memory_allocator/extern_vulkan_memory_allocator\nSPIRV-Tools-opt\nSPIRV-Tools\nSPIRV-Tools-link\nglslang\nSPIRV\nSPVRemapper/" -i source/blender/gpu/CMakeLists.txt || die
 	fi
 
-	rm "${WORKDIR}/blender-assets/publish/LICENSE" || die
+	if [ "${HAS_ASSETS}" -eq 1 ]; then
+		rm "${WORKDIR}/blender-assets/publish/LICENSE" || die
+	fi
 
-	sed -e "s/\"libhiprt64.so\"/\"libhiprt64.so.2.5\"/" -i extern/hipew/src/hiprtew.cc || die
+	sed -e "s/\"libhiprt64.so\"/\"libhiprt64.so.2.3\"/" -i extern/hipew/src/hiprtew.cc || die
 }
 
 src_configure() {
@@ -267,7 +305,7 @@ src_configure() {
 		-DBUILD_SHARED_LIBS=no
 		-DHIP_HIPCC_FLAGS="-fcf-protection=none"
 		-DHIP_LINKER_EXECUTABLE="$(get_llvm_prefix)/bin/clang++"
-		-DHIPRT_ROOT_DIR="/usr/include/hiprt/02005/"
+		-DHIPRT_ROOT_DIR="/usr/include/hiprt/02003/"
 		-DPYTHON_INCLUDE_DIR="$(python_get_includedir)"
 		-DPYTHON_LIBRARY="$(python_get_library_path)"
 		-DPYTHON_VERSION="${EPYTHON/python/}"
@@ -354,6 +392,12 @@ src_configure() {
 		-DWITH_PYTHON_MODULE=on
 	)
 
+	if use ffmpeg; then
+		ffmpeg_compat_setup 6
+		ffmpeg_compat_add_flags
+		mycmakeargs+=( -DFFMPEG_ROOT="${SYSROOT}$(ffmpeg_compat_get_prefix 6)" )
+	fi
+
 	if has_version ">=dev-python/numpy-2"; then
 		mycmakeargs+=(
 			-DPYTHON_NUMPY_INCLUDE_DIRS="$(python_get_sitedir)/numpy/_core/include"
@@ -398,16 +442,22 @@ src_configure() {
 src_install() {
 	cmake_src_install
 
+	if [ "${HAS_ASSETS}" -eq 1 ]; then
+		blender_get_version
+		insinto "$(python_get_sitedir)/bpy/${BV}/datafiles/assets"
+		doins -r "${WORKDIR}/blender-assets/publish/"*
+	fi
+
 	python_optimize "${D}$(python_get_sitedir)"
 }
 
 pkg_postinst() {
-	if ! use python_single_target_python3_13; then
+	if ! use python_single_target_python3_11; then
 		ewarn
 		ewarn "You are building Blender with a newer python version than"
 		ewarn "supported by this version upstream."
 		ewarn "If you experience breakages with e.g. plugins, please switch to"
-		ewarn "python_single_target_python3_13 instead."
+		ewarn "python_single_target_python3_11 instead."
 		ewarn "Bug: https://bugs.gentoo.org/737388"
 		ewarn
 	fi
